Make MGW nat more flexible by introdicng actors an function references
Instead of a static configuration of rewrite/mangling/patching steps,
we now introduce the concept of a 'rewrite_actor', which is a callback
function that determines which rewrite/patching code to call at which
particular layer of the protocol.

The default rewrite actor "bow_onw" resembles the existing rewrite
behavior.

There is a stub new rewrite actore "vfuk_onw", which will be used
for Camel phase rewriting.