add new mangle_tt_sri_sm module and associated eunit tests
This module is able to match a SCCP message against a given
Destination-GT prefix and checks if the message contains an
Invoke(SRI-for-SM).  If both conditions are true, it alters the TT from
whatever it may currently be to 3.

This is used for SMS special routing cases at the STP.