mangle_callbarr: Add code to manlge call barring in InsertSubscrData
When the HLR inserts subscriber data into the VLR, we want to mangle the
callBarringInfo, which is part of the provisionedSS member in the
InsertSubscriberDataArg type.

Instead of the original call barring list, a new list (generated from
the configuration file) is used.  The config file syntax for this is
liek this:
	{mangle_callbarr_list, [ {23,"PA"}, {42,"QR"} ] },

which configures two teleservices:
	* 23 with Privisioned+Active and
	* 42 with Q and Registered