mangle_callbarr: don't replace any callBarringInfo, but only BAOC
We want to make sure that any InsertSubscriberDataArg from HLR to VLR
always contains the BAOC as configued in the mgw_nat config file.

So what we do is as follows:

1) if the HLR ever sends BAOC, replace it with the BAOC generated
   from our config file
2) always after the provisionedSS(COLR), we unconditionally insert
   our BAOC

'2' is required as we don't have any state beyond the current message
    and don't know if there has been BAOC present in previous
    InsertSubscriberDataArg.

The decision to insert BAOC after COLR was made after analyzing the
actual output of the HLR in question, as the ISD with the BAOC was the
smallest ISD message, and we want to avoid increasing signalling unit
size to not hit MSU limits somewhere in the global SS7 networks.